Machine Learning/Model 2/requirements.txt
Flask==3.0.3
tensorflow==2.16.1
opencv-python-headless==4.5.5.64
matplotlib==3.5.1
numpy==1.26.4
rembg
pillow
colorthief
gunicorn